restructured som sorting function
