+++ src/js/sorter/sortType.mjs
@@ -1,5 +1,6 @@
 import { API_BASE_URL } from "../auth/constants.mjs";
 import { sortedPosts } from "../handlers/sortPost.mjs";
+import { displaySortInfo } from "./sortInfo.mjs";
 
 /**
  * // Function that retrieves queryString from the url of the current page it's on and passes it into a API endpoint fetch call, to sort and show the posts in the sorted order that's in the parameter value
@@ -17,16 +18,6 @@ export function sortPosts() {
   const sortOrder = params.get("sortOrder");
   const sortPath = `/social/posts?sort=${sortType}&sortOrder=${sortOrder}&_author=true&_reactions=true&_comments=true`;
   const API_SORT_URL = `${API_BASE_URL}${sortPath}`;
-
-  const sortInfo = document.querySelector(".type-of-sort");
-  if (sortType === "title" && sortOrder === "asc") {
-    sortInfo.innerText += ` Title A-Z`;
-  } else if (sortType === "title" && sortOrder === "desc") {
-    sortInfo.innerText += ` Title Z-A`;
-  } else if (sortType === "created" && sortOrder === "asc") {
-    sortInfo.innerText += ` Oldest Posts`;
-  } else if (sortType === "created" && sortOrder === "desc") {
-    sortInfo.innerText += ` Newest Posts`;
-  }
+  displaySortInfo(sortType, sortOrder);
   sortedPosts(API_SORT_URL);
 }
